GET: Fix confusion over the type of the 'type' parameter.
The 'type' parameter was being used to pass in an "enum reader_command" but
it was then later assigned a value of "enum pfm_type".  This renames 'type'
to 'command' for clarity and just deletes the assignment of the "enum
pfm_type" value entirely because the value was not used.

Found by Clang.

Reported-by: Jeremy Lavergne <jeremy@lavergne.gotdns.org>