gui: Speed up initial file load with many columns
This makes a huge difference in startup time when there are hundreds of
columns, by setting up the data sheet before showing the window that
contains it.

Files opened from File|Open were already opened this way.  I was a little
surprised that File|Open was so much faster than specifying a file on the
command line until I figured out the difference.