Output Viewer:  Try to adapt to changes in style
Set the fg and bg of viewer's container and its children to the
text and base of the output window upon style change.  This means that
the output viewer will respond to changes in style once even after the
program has been started.

We still need some way to set the foreground colours of the cairo output
after the driver has been created, but that doesn't seem to be currently
possible.