Name the src/libpspp convenience library consistently.
src/libpspp.la and src/libpspp/libpspp.la have the same name,
although they are in different directories.  This seems unduly
confusing.  This commit renames the latter to liblibpspp.la, which
is at least different from libpspp.la and is also consistent with
the naming of the other libtool convenience libraries as
lib<directory>.la.