Reliability Test: Do not allow splits greater than the number of variables.
It's not meaningful to have a split point which is greater than the number
of variables (and not useful when they are equal).  And the system was
exhaughsting its memory when such a test was run.  This change therefore
disallows such a combination and updates the behaviour of the GUI accordingly.