encoding-guesser: Fall back to windows-1252 when UTF-8 can't be right.
Until now the encoding-guesser code has used UTF-8 as a fallback in
situations where we can tell that the file is not valid UTF-8.  In
this kind of situation having a single-byte character set as a
fallback makes more sense.  This commit hard-codes windows-1252 as
that fallback, since it is a widely encountered encoding (and
compatible with ISO-8859-1 as well).

John Darrington originally suggested this, if I recall correctly.

The bug report that spurred this work was from Harry Thijssen.  With
this commit, PSPP properly reads his windows-1252 file when the
system locale uses UTF-8 encoding.