Don't update the icon cache if DESTDIR is non-empty.
Packagers generally use "make install DESTDIR=xxx" to generate their install
files, which are then copied in place to the target filesystem.  In such a
scenario, it's inappropriate to copy the icon.cache files.  This change
therefore does not update the icon cache if a DESTDIR has been specified.
Instead it's left to the packager to make his own arrangements for updating
the icon cache on the target filesystem.