sort: Allocate space for sort buffers on demand, not in advance.
Each call to sort_create_writer() was allocating possibly a very
large amount of memory when in many cases only a small fraction of
that memory would actually be used.  This commit switches to
allocating a small amount of memory and gradually increasing the
allocation on demand.

Reported by John Darrington.
Bug #35887.