PspppireSelector: Prevent selection of disallowed entries.
We had been relying on the deselected state of the widget to prevent
selection with the "allow" predicate was false.   But clicking the
widget itself isn't the only way items can be selected (double clicking
the item for example).   This change adds a test at a lower level to
prevent selection of disallowed items.