pspp-widget-facade: New code to measure and render some GTK+ widgets.
There are two upcoming uses for this in the tree, both related to
rendering fake buttons.  First, creating, maintaining, and rendering
the column header buttons in a PsppSheetView is a huge cost when
there are hundreds or thousands of columns.  Using this facade code
to "fake" buttons makes the sheet much faster.  Second, this code
will be used to implement a cell renderer for GtkButton, so that
buttons can be put into PsppSheetView rows.