gui: Properly manage ref counts of combo box and tree view models.
gtk_combo_box_set_model() and gtk_tree_view_set_model() add a
reference to the passed-in model, instead of transferring a
reference from the caller.  A lot of the calls to these functions,
however, assumed that a reference was transferred and therefore
leaked a reference.  This fixes them up.

This doesn't change PsppireValueEntry, because there's an ongoing
discussion about that.