psppire-data-window: Destroy dataset from finalize.
One rule of thumb for "destroy" is that it should just unref
GObjects.  This commit fixes that up for
psppire_data_window_dispose().  Previously, this could not be done
because of a reference leak in psppire_data_window_finish_init()
that caused doing the dataset destroy in finalize to crash, but
that problem was fixed in the previous commit.