psppire-var-sheet: Allow adjusting width of string variables.
John Darrington pointed out that the "width" column in the variable
sheet didn't allow the width to be modified with a spin button for
string variables, as it did for numeric variables.  This commit
fixes the problem.

The reordering in on_var_column_edited() is necessary because the
print format of a string variable can't be changed to a different
width before changing the variable's width.  (The change from
var_set_print_format() to var_set_both_formats() is not
necessary.  It is for consistency with changing variable type and
format via the var-type-dialog.)