gnumeric-reader: Add cast to suppress warning.
This suppresses a GCC warning for a mismatch between
xmlInputCloseCallback, which takes a "void *" parameter, and
gzclose(), which takes a "gzFile" (which is "struct gzFile_s *").

This came up with an upgrade from "squeeze" to "wheezy".  I assume
that a gzclose prototype became visible or that gzFile changed away
from "void *".

An alternative would be to write a trampoline to just transform one
pointer type to the other, but the xmlReaderForIO() invocation
already has one function cast so a second cast doesn't make it much
worse.