tests: Better isolate search text in test of error reporting.
Some GNU/Linux distributions print additional text on a segfault,
e.g. "Cannot exec debugger: No such file or directory", which this
test was seeing in output and interpreting as a failed test.  This
commit fixes the problem by ignoring text up to the ***** line that
marks the start of the expected output.

Bug #37444.
Reported by Mindaugus.