Fix bug where, after double-clicking to switch sheet view, a spurious release event was processed.
From http://developer.gnome.org/gdk/2.24/gdk-Event-Structures.html#GdkEventButton

  Double and triple-clicks result in a sequence of events being received. For
  double-clicks the order of events will be:

   1.  GDK_BUTTON_PRESS
   2.  GDK_BUTTON_RELEASE
   3.  GDK_BUTTON_PRESS
   4.  GDK_2BUTTON_PRESS
   5.  GDK_BUTTON_RELEASE

So what was happening was this:  When the var-name was double-clicked, the view switched
from VarView to DataView as soon as  event 4 (GDK_2BUTTON_PRESS) occured.  Then event 5
(GDK_BUTTON_RELEASE) occured.

Now the PsppSheetView object is written such that, it decides to edit a cell, on button
release event. Hence, whichever cell the pointer happended to be in, when the user lifts
his finger after double clicking, starts to be edited.

This change ignores button release events on the object unless a corresponding button press
event was seen.