gui: Make Tab, Shift+Tab skip row head and new-var columns in data sheet.
Before this commit, Tab and Shift+Tab acted just like Left and Right in
the data sheet, that is, they visit every column except the row head
column and always stay on the same row.  After this commit, Tab and
Shift+Tab skip the row head and new-variable columns.  Also, Tab at the
end of a row goes to the beginning of the next row, and Shift+Tab at the
beginning of a row goes to the end of the previous row.

Suggested by John Darrington.