pspp-sheet-view: Fix warning using arrow keys to navigate column buttons.
Before this commit, clicking on a column head button in the data sheet,
then holding down the right arrow key until the data sheet began to scroll
right would often produce a warning of the form:

** (psppire:16451): WARNING **: Internal button not found

because nothing caused the underlying GtkButton to be instantiated in this
scenario.  This commit fixes the problem.