psppire-button-editable: Fix infinite recursion in button-release-event.
"release" and "button-release-event" are different.  Before commit
8f843c3a5e905c (Avoid deprecated function: gtk_button_released),
psppire_button_editable_button_release() invoked the former.  That commit
changed it to invoke the latter, which is the same signal that the
function itself implements, hence causing infinite recursion.  This commit
changes it to invoke the former.

Without this commit, the following instructions reliably crash for me:

1. Load a data file, click over to variables tab.
2. Click on a ... button in the Type column.
3. Click on Cancel button in Type dialog.
4. Click on the same ... button as before.

With this commit, those instructions don't crash (the Type dialog pops
back up after step 4, as it should).

Reported by John Darrington.