DATA LIST FREE: Warn when a quoted string is not followed by a delimiter.
A field parsed by DATA LIST FREE (or LIST) is either quoted or unquoted.
When it is quoted, this means that the entire field has to be quoted, that
is, it is not possible to quote some initial part of the field, then leave
some part of it unquoted (then possibly quote an additional part, etc.).
That means that the quoted portion should be followed by a delimiter,
such as white space, a comma, or end of line, but PSPP didn't warn when
it was not.  This commit makes PSPP warn in this situation, which makes
errors in the input easier to find.

Bug #26522.
Reported by Pascal Barbedor <pbarbedor@gmail.com>.