data-parser: Use data-file error types for data-file errors.
The SW error type that these calls to msg() used before this command
include the name of the executing command in the error message.  This can
be confusing because the executing command may not have any connection to
the command that said to read the data file.  This commit changes these
error types to DW, which is intended for errors in a data file (and omits
the command name).

Bug #26517.