ascii: Omit whitespace at the end of a line.
Long string variables tend not to be completely filled with data, which
means that they end in a lot of spaces.  The ASCII output driver didn't
handle these well.  Instead, it put a blank line in the cell for each
trailing space.

This fixes the problem and adds a test.

Reported by Ronald Crichton.
Bug #38672.