Fix stack corruption in ODS reader.
zip_reader_create takes a pointer to a struct string, which
is used to  store error messages during zip operations.  I had
been passing the address of a string declared as local variable
in ods_probe.  However since the zip reader survives ods_probe,
this caused stack corruption in later read operations.

This change makes the string a member of the reader created (if any).