Avoid letting data fields overflow in output and GUI.
When the output for data_out() is too big for the field width, it produces
output that consists of just asterisks: ******.  This is OK when the
output is really going into a fixed-width space, such as the output for
PRINT and WRITE.  But it is obnoxious if the output is going into PSPP
output or the GUI.  This commit introduces a new function that typically
does not do this, and starts using it in output and the GUI.

Bug #35829.
Reported by John Darrington.
Bug #30731.
Reported by lavila <la@jump23.com>.