INPUT PROGRAM: Use a separate dataset for the input program.
Without this change, INPUT PROGRAM uses the same dataset as the following
syntax.  This usually is fine, but there's at least one case where it
doesn't work: if the INPUT PROGRAM defines a vector, then a following
proc_execute() without any intervening transformations causes
proc_execute() to follow a "fast path" where it skips running the procedure
but instead just does a few things, like clearing the set of vectors.
Unfortunately, this means that later when the input program is really run,
  it doesn't have the vectors and accesses freed memory.

This commit switches to a different tactic.  INPUT PROGRAM creates a
special session and dataset that are used only for the input program.  This
keeps the dataset and the transformations used for the input program
separate from those used after the input program.

Bug #39097.
Reported by John Darrington.