datasheet: Avoid reading a given row more than necessary in source_read().
The datasheet casereader uses datasheet_get_row(), which calls rw_case(),
which until this commit called source_read() for every column.  The latter
was expensive if casereader_peek() actually copied data for a whole large
row.  For a datasheet with N columns, all from a single backing source,
the upshot was that datasheet_get_row() took O(N**2) time.

This commit fixes the problem in the common case by reading as many columns
as possible from a given source in a single call to source_read().

Bug #33260.