gtkxpaned: Remove style code from gtk_xpaned_realize().
gtk_style_attach() is deprecated in GTK+3 and the migration guide says
that it is no longer needed.  I guess it is possible that the
gtk_widget_set_style() call is somehow still needed, but I don't
understand the intent of that call and removing it seems to have no
visible effect.