Make one of the icon sets a wildcard set.
We currently have two icon sets: A 16x16 and a 24x24 set.  Previously,
we had been using the 16x16 set for menuitems and the 24x24 set for the
toolbar.  However it appeared that windows uses the symbol GTK_ICON_SIZE_SMALL_TOOLBAR
instead of GTK_ICON_SIZE_LARGE_TOOLBAR, which meant no icons appeared in the
toolbar.

This change sets the 24x24 set as a "wildcard" set meaning it will be used for
all purposes except menuitems.