variable: Only notify variable width change if it really did change.
This code was notifying a width change every time, even if the width did
not really change, which is wasteful.  This had a side effect of revealing
a bug in the GUI code which caused an assertion failure.  The bug still
remains in the GUI code, though it is now slightly harder to trigger, and
will be fixed separately in an upcoming commit.