Replace checkbox-treeview functions with a object PsppireCheckboxTreeview
Prior to this change there existed a function which operated on
a GtkTreeView which populated it with a model and renderers such
that it would display a list of annotated checkboxes.

This change subclasses GtkTreeView and moves the function to a
method of that class.

I anticipate this will ease implementation of upcoming functionality.