comments-dialog: Avoid segfault in comments dialog pasting very long text.
The code in wrap_line() used a GtkTextIter after an insertion invalidated
it.  In some setups, with a sufficiently long pasted text, this caused a
Gtk-WARNING and a subsequent segfault.  This commit fixes the problem by
removing the offending code.  In my testing, this didn't cause a UI
regression.

Reported by Bastián Díaz in the thread starting at
http://lists.gnu.org/archive/html/bug-gnu-pspp/2013-10/msg00005.html