Respect forground and background options when writing PNG charts.
The default colours for charts in PNG format was black on black,
which wasn't very useful.  This change makes the defaults black
on white.

Perhaps for png charts intended for embedding in HTML it should be
black on transparent, but this hasn't been implemented in this change.

Fixes bugs #40107 and #40208