i18n: Make recode_substring_pool() always return a pool_malloc()'d string.
recode_substring_pool() generally returned a string allocated by
pool_malloc(), but when there was an error finding the converter it
returned a string that was allocated by pool_alloc().  This was a problem
for parse_value_labels() in the system file reader, because that function
frees all the strings it recodes.  This commit fixes the problem by
always using pool_malloc() in recode_substring_pool().